gt - Correct labels when no country
Update stories
