+++ stories/index.js
@@ -28,7 +28,6 @@ import './ImageCarousel';
 import './Input';
 import './InfoBox';
 import './InputGroup';
-import './InternationalAddressInput';
 import './LabelBadge';
 import './Modal';
 import './MonthInput';
